[GR-37381] [GR-35988] Consolidate NativeImageClassLoaderSupport.
PullRequest: graal/11359
