+++ substratevm/src/com.oracle.svm.test/src/META-INF/native-image/com.oracle.svm.test/native-image.properties
@@ -5,6 +5,7 @@ Args= \
   --features=com.oracle.svm.test.AbstractServiceLoaderTest$TestFeature \
   --features=com.oracle.svm.test.NoProviderConstructorServiceLoaderTest$TestFeature \
   --features=com.oracle.svm.test.NativeImageResourceUtils$TestFeature \
+  --features=com.oracle.svm.test.jfr.JFRTestFeature \
   --add-opens=java.base/java.lang=ALL-UNNAMED \
   -H:+AllowVMInspection \
   --add-exports=org.graalvm.nativeimage.builder/com.oracle.svm.core.containers=ALL-UNNAMED
